refactor: cutting file conversion to checkarray and change dim, file check move in module check, change structure change dim with args just array
