+++ .github/workflows/lint_python.yml
@@ -36,7 +36,7 @@ jobs:
         pylint src/reader/orientation/manage_reader.py src/reader/orientation/reader_opk.py
         pylint src/datastruct/shot.py src/datastruct/gcp.py src/datastruct/worksite.py src/datastruct/camera.py src/datastruct/dtm.py
         pylint src/writer/writer_opk.py src/writer/manage_writer.py
-        pylint src/utils/conversion.py src/utils/check_header.py src/utils/singleton/singleton.py
+        pylint src/utils/change_dim.py src/utils/check/check_header.py src/utils/check/check_array.py src/utils/singleton/singleton.py
         pylint src/transform_world_image/transform_worksite/space_resection.py src/transform_world_image/transform_worksite/copoints_pos.py
         pylint src/transform_world_image/transform_shot/world_image_shot.py src/transform_world_image/transform_shot/image_world_shot.py src/transform_world_image/transform_shot/conversion_coor_shot.py
         pylint src/transform_world_image/transform_dtm/world_image_dtm.py
@@ -51,7 +51,7 @@ jobs:
         flake8  --max-line-length 100 src/reader/orientation/manage_reader.py src/reader/orientation/reader_opk.py
         flake8  --max-line-length 100 src/datastruct/shot.py src/datastruct/gcp.py src/datastruct/worksite.py src/datastruct/camera.py src/datastruct/dtm.py
         flake8  --max-line-length 100 src/writer/writer_opk.py src/writer/manage_writer.py
-        flake8  --max-line-length 100 src/utils/conversion.py src/utils/check_header.py src/utils/singleton/singleton.py
+        flake8  --max-line-length 100 src/utils/change_dim.py src/utils/check/check_header.py src/utils/check/check_array.py src/utils/singleton/singleton.py
         flake8  --max-line-length 100 src/transform_world_image/transform_worksite/space_resection.py src/transform_world_image/transform_worksite/copoints_pos.py
         flake8  --max-line-length 100 src/transform_world_image/transform_shot/world_image_shot.py src/transform_world_image/transform_shot/image_world_shot.py src/transform_world_image/transform_shot/conversion_coor_shot.py
         flake8  --max-line-length 100 src/transform_world_image/transform_dtm/world_image_dtm.py
