+++ ci/test_changedim.sh
@@ -1,0 +1,1 @@
+python3 -m pytest -s ./test/test_utils/test_changedim.py -v
